Create an addVegaReparseRequest method to replace the old parse mixin
Mixins are semi-unrecommended because they are not supported by React's
ES6 class syntax; we were only using the parse mixin to add what now
amounts to an action dispatcher to a view instance, so while we could
(and until now did) use `connect` to accomplish the same thing, we can
augment the module with less overhead.

We may need to switch back to Connect for testing purposes later on
but it was breaking more tests than it fixed for now.